Handle Docker push credentials more securely
Currently, we load the docker configuration in `.docker`, which *could*
(but hasn't in any of our repos) end up in the Docker image if we ran a
`ADD . somewhere` instruction.

This updates the Travis build file to copy the Docker configuration
outside of the build directory so we can't accidentally `ADD` it in the
Docker image.